transcription/requirements.txt
openai-whisper
rich
yt-dlp
halo
